pyproject.toml
[project]
name = "pdfmark"
version = "0.1.0"
description = "A tool to add beautiful gradient watermarks to PDF files using wave patterns and rainbow colors"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "numpy>=2.3.0",
    "pillow>=11.2.1",
    "pymupdf>=1.26.0",
    "pypdf2>=3.0.1",
    "reportlab>=4.4.1",
    "typer>=0.16.0",
]

[project.scripts]
pdfmark = "main:app"
